backend/requirements.txt
ansible-core
ansible-runner
bcrypt 
Django
djangorestframework
django-tenants
django-cors-headers
psycopg2-binary
celery
django-celery-beat
django_celery_results  
python-dateutil
python-dotenv 
pyjwt
djangorestframework-simplejwt
cryptography
azure-storage-blob
elasticsearch
python-dotenv
gunicorn
django-filter
python-magic
django-storages
django-extensions
setuptools
python-dotenv
django-redis
drf-yasg
docker
paramiko
